+++ test/client.js
@@ -14,16 +14,18 @@ describe('Test mock', () => {
     mock.stop(); // Stop Mock server
   });
 
+  beforeEach(() => {
+    return mock.get('/users/test', { hello: 'world' }).reply(200, {
+      id: 'test-id',
+      username: 'username',
+    });
+  });
+
   afterEach(() => {
     return mock.clean(); // Clean state after each test
   });
 
   it('should mock a user request', () => {
-    mock.get('/users/test').reply(200, {
-      id: 'test-id',
-      username: 'username',
-    });
-
     const options = {
       method: 'GET',
       uri: `${mock.url}/users/test`,
@@ -32,6 +34,7 @@ describe('Test mock', () => {
       },
       json: true,
       resolveWithFullResponse: true,
+      body: { hello: 'world' },
     };
     return rp(options).then((res) => {
       expect(res.statusCode).to.equal(200);
@@ -40,15 +43,20 @@ describe('Test mock', () => {
     });
   });
 
-  it('should issue an error when calling a function on finished request', () => {
-    const func = () => {
-      mock.get('/users/test')
-        .reply(200, {
-          id: 'test-id',
-          username: 'username',
-        })
-        .repeat(100);
-    };
-    expect(func).to.throw(/may not be called after reply/);
+  it('should get running test', () => {
+    return mock.getTest().then((test) => {
+      expect(test.expectations.length).to.equal(1);
+      expect(test.expectations[0].request.method).to.equal('get');
+      expect(test.expectations[0].request.url).to.equal('/users/test');
+      expect(test.expectations[0].request.body).to.deep.equal({ hello: 'world' });
+      expect(test.expectations[0].response.status).to.equal(200);
+      expect(test.expectations[0].response.body).to.deep.equal({
+        id: 'test-id',
+        username: 'username',
+      });
+      expect(test.expectations[0].repeat).to.equal(1);
+      expect(test.expectations[0].testId).to.equal('e2e');
+      expect(test.expectations[0].id.length).to.equal(36);
+    })
   });
 });