+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## [Unreleased]
 ### Changed
 - Prepublish now also runs test suite
+- Refactor `.reply()` to return promise
+- Rename `.end()` to `._end()` to indicate it is a private method
+- Add test for `mockingbird.getTest()`
 
 ## [0.0.3] - 2016-10-24
 ### Changed